Point the exporters at the moved M-RoPE layouts
